update(extensions_3d_tiles_temporal):new class STLayer
