+++ packages/extensions_3d_tiles_temporal/src/STSVector.js
@@ -1,3 +1,16 @@
 import { STShape } from './STShape';
 
-export class STSVector extends STShape {}
+export class STSVector extends STShape {
+  constructor(stLayer, options = {}) {
+    super(stLayer);
+  }
+  display() {
+    super.display();
+  }
+
+  update() {}
+
+  dispose() {
+    super.dispose();
+  }
+}
